feat: headset auto switch

+++ src/usb.h
@@ -6,6 +6,6 @@
 #define DS5_BRIDGE_USB_H
 
 extern uint8_t mute[2]; // 0: SPEAKER(0x02) 1: MIC(0x05)
-extern int16_t volume[2]; // 0: SPEAKER(0x02) 1: MIC(0x05)
+extern float volume[2]; // 0: SPEAKER(0x02) 1: MIC(0x05)
 
 #endif //DS5_BRIDGE_USB_H
